Added helper function to check user exixts

+++ user_functions/delete_user.py
@@ -2,7 +2,17 @@
 # Author ID: Mohammadullah
 
 import os
+import pwd
 
+def check_user_exists(username):
+    """This is a helper funtions that checks if username is valid."""
+    all_user = pwd.getpwall() 
+    for user in all_user:
+        if user.pw_name == username:
+            return True
+    return False
+
+    
 def delete_user():
     """This function deletes user from the system.
      It will ask the user the username that they intended to delete then uses os.system() to run the commands. """
@@ -29,12 +39,15 @@ def delete_user():
      # Step 4: OS command for deleting the user
     # userdel -r removes the user and user's home directory
     # the returned value of os.system is the key to decide whether the command in the string parameter is successfully executed or not. 
-    result = os.system(f"sudo userdel -r {username}")
+    if check_user_exists(username):
+
+        result = os.system(f"sudo userdel -r {username}")
+        print(f"User '{username}' successfully deleted.")
 
 
     # Step 5: Run the command to see what it returns
     if result == 0: # If the return value is 0 then it is showing that the task is successfully executed.
-        print(f"User '{username}' successfully deleted.")
+        
     else: # If the return value is not equal to zero, then the command failed
         # Reasons for failure may include:
         # - User does not exist
